Renamed validators
